created functional filter for types of wine
